src/main/java/io/builders/iobank/domain/service/impl/UserServiceImpl.java
package io.builders.iobank.domain.service.impl;

import io.builders.iobank.domain.exceptions.UserNotFoundException;
import io.builders.iobank.domain.model.User;
import io.builders.iobank.domain.port.repository.database.UserRepository;
import io.builders.iobank.domain.service.UserService;
import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.stereotype.Service;

import java.util.List;

@RequiredArgsConstructor
@Slf4j
@Service
public class UserServiceImpl implements UserService {

    private final UserRepository userRepository;

    @Override
    public User getUser(Long id) {
        return userRepository.findById(id)
                .orElseThrow(() -> new UserNotFoundException("User with id " + id + " not found"));
    }

    @Override
    public List<User> getUsers() {
        return userRepository.findAll();
    }

    @Override
    public User saveUser(User user) {
        return userRepository.save(user);
    }
}
